Fix ManyToMany Visita/Visitante + tests

+++ src/main/java/br/unesp/springcondominio/repository/VisitaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.unesp.springcondominio.entity.Visita;
 
 @Repository
-public interface VisitaRepository extends JpaRepository<Visita, Integer>{
+public interface VisitaRepository extends JpaRepository<Visita, Long>{
 
 }
